chore: refactor to move examples into separate crates
This makes isolating test cases easier.
