+++ cmd/gravity/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+authors = ["The Arcjet team"]
+categories = ["wasm"]
+default-run = "gravity"
+name = "arcjet-gravity"
+version = "0.0.2"
+edition = "2024"
+license = "Apache-2.0"
+repository = "https://github.com/arcjet/gravity"
+description = """
+Gravity is a host generator for WebAssembly Components. It currently targets Wazero, a zero dependency WebAssembly runtime for Go.
+"""
+
+[[bin]]
+name = "gravity"
+path = "src/main.rs"
+
+[dependencies]
+clap = "=4.5.36"
+genco = "=0.17.10"
+wit-bindgen-core = "=0.39.0"
+wit-component = "=0.225.0"
+
+[dev-dependencies]
+wit-bindgen = "=0.39.0"
