+++ examples/basic/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "example-basic"
+version = "0.1.0"
+edition = "2024"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+wit-bindgen = "0.42.1"
+wit-component = "0.230.0"
